- Fixed skill macro bug stemming from, you guessed it, more datatype issues!
git-svn-id: svn://70.113.44.214/VanaMirror/trunk@1354 09c1253f-3d09-8e40-9719-4320e56601b5

+++ src/ChannelServer/PlayerPacket.cpp
@@ -167,7 +167,7 @@ void PlayerPacket::showSkillMacros(Player *player, SkillMacros *macros) {
 	PacketCreator packet;
 	packet.addShort(SEND_SKILL_MACRO);
 	packet.addByte(macros->getMax() + 1);
-	for (int32_t i = 0; i <= macros->getMax();  i++) {
+	for (int8_t i = 0; i <= macros->getMax();  i++) {
 		SkillMacros::SkillMacro *macro = macros->getSkillMacro(i);
 		if (macro != 0) {
 			packet.addString(macro->name);
